Move discussion_api to discussion/rest_api
