+++ lms/djangoapps/discussion/rest_api/exceptions.py
