+++ lms/djangoapps/discussion/rest_api/tests/test_forms.py
@@ -9,7 +9,7 @@ import ddt
 from django.http import QueryDict
 from opaque_keys.edx.locator import CourseLocator
 
-from discussion_api.forms import CommentListGetForm, ThreadListGetForm
+from lms.djangoapps.discussion.rest_api.forms import CommentListGetForm, ThreadListGetForm
 from openedx.core.djangoapps.util.test_forms import FormTestMixin
