+++ lms/djangoapps/discussion/rest_api/tests/test_serializers.py
@@ -9,8 +9,8 @@ import httpretty
 import mock
 from django.test.client import RequestFactory
 
-from discussion_api.serializers import CommentSerializer, ThreadSerializer, get_context
-from discussion_api.tests.utils import CommentsServiceMockMixin, make_minimal_cs_comment, make_minimal_cs_thread
+from lms.djangoapps.discussion.rest_api.serializers import CommentSerializer, ThreadSerializer, get_context
+from lms.djangoapps.discussion.rest_api.tests.utils import CommentsServiceMockMixin, make_minimal_cs_comment, make_minimal_cs_thread
 from django_comment_client.tests.utils import ForumsEnableMixin
 from django_comment_common.comment_client.comment import Comment
 from django_comment_common.comment_client.thread import Thread
