--- lms/djangoapps/discussion_api/README.rst
@@ -1,1 +1,0 @@
-See ``lms/djangoapps/discussion/README.rst``
