+++ lms/envs/common.py
@@ -2125,7 +2125,6 @@ INSTALLED_APPS = [
     # Discussion forums
     'django_comment_client',
     'django_comment_common',
-    'discussion_api',
 
     # Notes
     'notes',
